my-application/requirements.txt
# Web framework
Flask==2.3.2
gunicorn==20.1.0

# Database driver
psycopg2-binary==2.9.5 # For connecting to Redshift

# Datadog APM (if using Datadog APM)
ddtrace==1.8.0

# Other common libraries (add as needed)
requests==2.31.0 # For making HTTP requests
python-dotenv==1.0.0 # For loading environment variables from .env files - for local development
psycopg2-binary==2.9.6 # Or your preferred version

# For local development and testing
pytest==7.0.0 # Or unittest

#For AI model
nltk==3.8.1